% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsObject_informatics.r
\name{AsObject}
\alias{AsObject}
\title{AsObject}
\usage{
AsObject(x)
}
\arguments{
\item{x}{Expression to be evaluated}
}
\value{
R object
}
\description{
Function to convert character string to object ('as object')
}
