% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/CreateDummyVarsFromMultChoiceMultAnsResponseStrings_informatics_2dimensions.r
\name{CreateDummyVarsFromMultChoiceMultAnsResponseStrings}
\alias{CreateDummyVarsFromMultChoiceMultAnsResponseStrings}
\title{CreateDummyVarsFromMultChoiceMultAnsResponseStrings}
\usage{
CreateDummyVarsFromMultChoiceMultAnsResponseStrings(
  tb,
  varname,
  answer.options,
  output.format = c("logical", "binary")
)
}
\arguments{
\item{tb}{Input table}

\item{varname}{Name of variable with multiple-choice multiple-answer response strings}

\item{answer.options}{Full list of unique answer options as a character vector (sometimes there will be an answer option that no respondent selected)}

\item{output.format}{"logical" for TRUE/FALSE, "binary" for 1/0 (default is 'logical')}
}
\value{
Original table containing input table and new variables.
}
\description{
Create dummy variables from single variable with concatenated strings containing responses to multiple-choice, multiple answer columns
}
