% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelLoadAllWorksheets_import.r
\name{ExcelLoadAllWorksheets}
\alias{ExcelLoadAllWorksheets}
\title{ExcelLoadAllWorksheets}
\usage{
ExcelLoadAllWorksheets(filename, tibble = TRUE)
}
\arguments{
\item{filename}{Name of Excel file to be imported}

\item{tibble}{If TRUE (default) returns tibbles, if FALSE, returns data frames.}
}
\value{
Set of tibbles or data frames with same names as the tab names in the imported Excel file.
}
\description{
Load all worksheets from an Excel File. Each sheet becomes an R object with the name of the sheet.
}
