% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoogleSheetLoadAllWorksheets_setup.r
\name{GoogleSheetLoadAllWorksheets}
\alias{GoogleSheetLoadAllWorksheets}
\title{GoogleSheetLoadAllWorksheets
Load all worksheets from a Google Sheet Object
Returns TRUE if expression throws an error, FALSE otherwise.}
\usage{
GoogleSheetLoadAllWorksheets(gs)
}
\arguments{
\item{gs}{Google Sheet object}
}
\description{
GoogleSheetLoadAllWorksheets
Load all worksheets from a Google Sheet Object
Returns TRUE if expression throws an error, FALSE otherwise.
}
