% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NotIn_operator.r
\name{\%!in\%}
\alias{\%!in\%}
\title{\%!in\%}
\usage{
x \%!in\% y
}
\description{
New operator that complements \%in\%. Whereas a \%in\% b returns the elements of a that are also in b, a \%!in\% b returns the elements of a that are NOT in b (a subtracting the intersection of a+b).
}
