% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/IndexMatchToVectorFromTibble_informatics_2dimensions.r
\name{IndexMatchToVectorFromTibble}
\alias{IndexMatchToVectorFromTibble}
\title{IndexMatchToVectorFromTibble}
\usage{
IndexMatchToVectorFromTibble(
  vector,
  lookup.tb,
  match.varname,
  replacement.vals.varname,
  mult.replacements.per.cell = c(FALSE, TRUE),
  mult.replacements.separator.char = NULL,
  print.matches = c(TRUE, FALSE)
)
}
\arguments{
\item{vector}{Vector with values to be replaced.}

\item{lookup.tb}{The lookup table with lookup values and corresponding replacements values}

\item{match.varname}{Name of variable in lookup table with values that match vector values to be replaced.}

\item{replacement.vals.varname}{Name of variable in the lookup table with values that will be substituted into the vector.}

\item{mult.replacements.per.cell}{Logical - whether there are multiple replacements in a single element of the vector.}

\item{mult.replacements.separator.char}{If mult.replacements.per.cell is TRUE, the character upon which elements are to be split before substitution.}

\item{print.matches}{Logical - whether to print a table of matches.}
}
\value{
R object
}
\description{
Index-match equivalent for replacing values of a vector with values from a tibble given variable name in table
}
