% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstallCommonPackages_setup.r
\name{InstallCommonPackages}
\alias{InstallCommonPackages}
\title{InstallCommonPackages}
\usage{
InstallCommonPackages()
}
\description{
Single command to install a number of commonly used packages.
Current list:
chron
devtools
readr
data.table
dplyr
tidyr
googlesheets
stringr
officer
magrittr
reshape2
ggplot2
xlsx
styler
}
