% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsError_errors.r
\name{IsError}
\alias{IsError}
\title{IsError}
\usage{
IsError(.expr)
}
\arguments{
\item{.expr}{Expression to be evaluated}
}
\value{
Logical
}
\description{
Returns TRUE if expression throws an error, FALSE otherwise.
}
