% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsSpecialChar_informatics_1dimension.r
\name{IsSpecialChar}
\alias{IsSpecialChar}
\title{IsSpecialChar}
\usage{
IsSpecialChar(char)
}
\arguments{
\item{char}{Character string input of length 1.}
}
\value{
Logical TRUE/FALSE
}
\description{
Detect Special Character and return TRUE/FALSE
}
