% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListToTibbleObjects_informatics_2dimensions.r
\name{ListToTibbleObjects}
\alias{ListToTibbleObjects}
\title{ListToTibbleObjects}
\usage{
ListToTibbleObjects(list)
}
\arguments{
\item{list}{List to be split out into tibble objects}
}
\value{
Set of tibbles with contents of each list element.
}
\description{
Converts all elements in a list to separate R objects (tibbles). Uses names of list elements to assign names to R Objects.
}
