% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_LoadCommonPackages.R
\name{LoadCommonPackages}
\alias{LoadCommonPackages}
\title{LoadCommon Packages}
\usage{
LoadCommonPackages()
}
\description{
Load commonly used packages:
data.table
ggplot2
googlesheets
magrittr
reshape2
stringr
tictoc
tidyverse
XLConnect
}
\details{
Function to load most commonly used packages.
}
