% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowerCaseCharVars_informatics_2dimensions.r
\name{LowerCaseCharVars}
\alias{LowerCaseCharVars}
\title{LowerCaseCharVars}
\usage{
LowerCaseCharVars(tb)
}
\arguments{
\item{tb}{Input table.}
}
\value{
Table with all character variables in lower-case.
}
\description{
Lower-case all character values in a tibble
}
