% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowerCaseNames_informatics_2dimensions.r
\name{LowerCaseNames}
\alias{LowerCaseNames}
\title{LowerCaseNames}
\usage{
LowerCaseNames(tb)
}
\arguments{
\item{tb}{Input table}
}
\value{
Table with names in lower-case
}
\description{
Lower-case names of a table.
}
