% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mgsub_informatics_1dimension.r
\name{Mgsub}
\alias{Mgsub}
\title{Mgsub}
\usage{
Mgsub(pattern, replacement, x)
}
\arguments{
\item{pattern}{Vector of character strings to be found and substituted.}

\item{replacement}{Vector of character strings to be put in place of pattern values. Length and order matters: replacement vector must be of the same length as pattern vector. Values correspond via order of vector (first element in replacement vector will be substituted for the first element of the pattern).}

\item{x}{Vector in which values are to be substituted.}
}
\value{
R object
}
\description{
'Multiple gsub' to find/replace multiple patterns in a vector
}
