% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_MostRecentlyModifiedFilename.R
\name{MostRecentlyModifiedFilename}
\alias{MostRecentlyModifiedFilename}
\title{MostRecentlyModifiedFilename}
\usage{
MostRecentlyModifiedFilename(title.string.match, file.type, dir)
}
\arguments{
\item{title.string.match}{Character string to be founding within the filename.}

\item{file.type}{File type extension.}

\item{dir}{Directory to search. This function only operates on a single directory and does not search subdirectories.}
}
\value{
Character string object for most recently modified file name
}
\description{
Find most recently modified file in a directory
}
