% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoveColsLeft_informatics_2dimensions.r
\name{MoveColsLeft}
\alias{MoveColsLeft}
\title{MoveColsLeft}
\usage{
MoveColsLeft(tb, varnames)
}
\arguments{
\item{tb}{Table in which column(s) will be moved.}

\item{varnames}{Names of variables to be moved.}
}
\value{
Tibble object
}
\description{
Move named columns to the front/left side of table
}
