% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoveColsRight_informatics_2dimensions.r
\name{MoveColsRight}
\alias{MoveColsRight}
\title{MoveColsRight}
\usage{
MoveColsRight(tb, varnames)
}
\arguments{
\item{tb}{Table in which column(s) will be moved.}

\item{varnames}{Names of variables to be moved.}
}
\value{
Tibble object
}
\description{
Move named columns to the back/right side of table
}
