% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumSubstringMatches_informatics_1dimension.r
\name{NumSubstringMatches}
\alias{NumSubstringMatches}
\title{NumSubstringMatches}
\usage{
NumSubstringMatches(substring, string)
}
\arguments{
\item{substring}{Substring pattern to be found within the string}

\item{string}{String to be searched}
}
\value{
Numeric object with number of matches for 'substring' within 'string'
}
\description{
Number of times specified substring occurs within a character string
}
