% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/NumSubstringMatchesInVector_informatics_1dimension.r
\name{NumSubstringMatchesInVector}
\alias{NumSubstringMatchesInVector}
\title{NumSubstringMatchesInVector}
\usage{
NumSubstringMatchesInVector(substring, vector)
}
\arguments{
\item{substring}{Substring pattern to be found within the each element of 'vector'}

\item{vector}{Vector to be searched}
}
\value{
Numeric vector object with number of matches for 'substring' within each element of 'vector'
}
\description{
Number of times specified substring occurs within a character string
}
