% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderTbByVar_informatics_2dimensions.r
\name{OrderTbByVar}
\alias{OrderTbByVar}
\title{OrderTbByVar}
\usage{
OrderTbByVar(tb, order.by.varname)
}
\arguments{
\item{tb}{}

\item{order.by.varname}{Character string indicating name of column by which table is to be ordered}
}
\value{
Ordered tibble
}
\description{
Order table by specific variable using the variable name
}
