% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintLoopNum.r
\name{PrintLoopNum}
\alias{PrintLoopNum}
\title{PrintLoopNum}
\usage{
PrintLoopNum(loop.index)
}
\arguments{
\item{loop.index}{Name of object using as the loop index}
}
\description{
When placed inside a loop, will print message like the following: "LOOP 'i' -- Loop num: 7"
}
