% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proper_informatics_strings.r
\name{Proper}
\alias{Proper}
\title{Proper}
\usage{
Proper(s, strict = FALSE)
}
\arguments{
\item{x}{Character (if multiple words, words must be separated by space)}
}
\value{
Returns x with the first letter of each word capitalized
}
\description{
Excel 'proper' function - capitalize first letter of each word in string
}
