% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/RemoveAllBlankNANullColumns_informatics_2dimensions.r
\name{RemoveAllBlankNAColumns}
\alias{RemoveAllBlankNAColumns}
\title{RemoveAllBlankNAColumns}
\usage{
RemoveAllBlankNAColumns(tb)
}
\arguments{
\item{tb}{Input table}
}
\value{
Table without columns where all rows were blank or NA
}
\description{
Select variables based on names in/not in a string vector
}
