% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveNA_informatics_1dimension.r
\name{RemoveNA}
\alias{RemoveNA}
\title{RemoveNA}
\usage{
RemoveNA(x)
}
\arguments{
\item{x}{Vector input from which all NAs will be removed.}
}
\value{
Vector with no NAs.
}
\description{
Removes NA from vector
}
