% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceNames_informatics_2dimensions.r
\name{ReplaceNames}
\alias{ReplaceNames}
\title{ReplaceNames}
\usage{
ReplaceNames(tb, current.names, new.names)
}
\arguments{
\item{tb}{Table where names will be replaced. Must be data frame or tibble.}

\item{current.names}{Names to be replaced. Must be a character string.}

\item{new.names}{New names that will be inserted in the place of current.names. Must be a character string.}
}
\value{
R data table or tibble object with different column names (replaces R object with original names)
}
\description{
Replace selected names the names in a data frame or tibble.
}
