% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/ReplaceRepeatingSpacesWithSingleSpace_informatics_strings.r
\name{ReplaceRepeatingSpacesWithSingleSpace}
\alias{ReplaceRepeatingSpacesWithSingleSpace}
\title{ReplaceRepeatingSpacesWithSingleSpace}
\usage{
ReplaceRepeatingSpacesWithSingleSpace(x)
}
\arguments{
\item{x}{Character string in which repeated spaces are to be replaced with a single space}
}
\value{
x with no repeated spaces
}
\description{
Reduce repeating substring with single occurrence of that substring
}
