% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/ReplaceRepeatingSubstringWithSingleSubstring_informatics_strings.r
\name{ReplaceRepeatingSubstringWithSingleSubstring}
\alias{ReplaceRepeatingSubstringWithSingleSubstring}
\title{ReplaceRepeatingSubstringWithSingleSubstring}
\usage{
ReplaceRepeatingSubstringWithSingleSubstring(substring, string)
}
\arguments{
\item{substring}{Repeated character pattern to be replaced with a single version of itself within 'string'}

\item{string}{Character string in which substrings are to be replaced}
}
\value{
'string' with no repeated substrings
}
\description{
Reduce repeating substring with single occurrence of that substring
}
