% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectVarsByName_informatics_2dimensions.r
\name{SelectVarsByName}
\alias{SelectVarsByName}
\title{SelectVarsByName}
\usage{
SelectVarsByName(tb, condition = c("IN", "NOT.IN"), char.vector)
}
\arguments{
\item{tb}{Input table}

\item{condition}{Either "IN" or "NOT.IN"}

\item{char.vector}{Character vector of names to be selected or excluded.}
}
\value{
Table
}
\description{
Select variables based on names in/not in a string vector
}
