% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SplitColReshape.ToLong_informatics_2dimensions.r
\name{SplitColReshape.ToLong}
\alias{SplitColReshape.ToLong}
\title{SplitColReshape.ToLong}
\usage{
SplitColReshape.ToLong(tb, id.varname, split.varname, split.char)
}
\arguments{
\item{tb}{Input table}

\item{id.varname}{Name of variable in input table to be used as an id/primary key.}

\item{split.varname}{Name of variable to be split and reshaped.}

\item{split.char}{Character designating where split variable should be split.}
}
\value{
Tibble
}
\description{
Reshape data from wide to long format based on splitting a column on a character
}
