% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubNA_informatics_1dimension.r
\name{SubNA}
\alias{SubNA}
\title{SubNA}
\usage{
SubNA(x, na.replacement)
}
\arguments{
\item{x}{Vector input. If no NAs in the input, function will return original input.}

\item{na.replacement}{Value with which NAs will be substituted.}
}
\value{
Vector
}
\description{
Replace NAs in a vector with a replacement value
}
