% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubstrRight_informatics.r
\name{SubstrRight}
\alias{SubstrRight}
\title{SubstrRight}
\usage{
SubstrRight(x, n)
}
\arguments{
\item{x}{Character string}

\item{n}{Number of characters to be selected, starting on the right end of x}
}
\value{
Character string object with the 'n' rightmost characters of 'x'
}
\description{
Select right 'n' characters of string
}
