% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableWithNA_informatics_1dimension.r
\name{TableWithNA}
\alias{TableWithNA}
\title{TableWithNA}
\usage{
TableWithNA(x)
}
\arguments{
\item{x}{Vector or variable input.}
}
\value{
R Table
}
\description{
Shortcut to use base R table() function and return number of NAs.
}
