% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TibbleToCharObjects_import.r
\name{TibbleToCharObjects}
\alias{TibbleToCharObjects}
\title{TibbleToCharObjects}
\usage{
TibbleToCharObjects(tibble, object.names.colname, object.values.colname)
}
\arguments{
\item{tibble}{Input tibble.}

\item{object.names.colname}{Name of column from which names for new objects to be created will be read.}

\item{object.values.colname}{Name of column from which values of the new objects will be created.}
}
\value{
Set of named R character string objects.
}
\description{
Assign values in column of a tibble to character string objects with name from the same row in another column.
}
