% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransposeTable_informatics_2dimensions.r
\name{TransposeTable}
\alias{TransposeTable}
\title{TransposeTable}
\usage{
TransposeTable(tb, keep.first.varname = c(TRUE, FALSE))
}
\arguments{
\item{tb}{Input table}

\item{keep.first.varname}{Logical TRUE/FALSE indicating whether to preserve the name of the first variable in the transposed table}
}
\value{
Table
}
\description{
Transpose table with correct column & row names.
}
