% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/UniqueCombnFromVarnames_informatics_2dimensions.r
\name{UniqueCombnFromVarnames}
\alias{UniqueCombnFromVarnames}
\title{UniqueCombnFromVarnames}
\usage{
UniqueCombnFromVarnames(tb, varnames)
}
\arguments{
\item{tb}{Table input with named variables.}

\item{varnames}{vector of variable names from which combinations of unique values will be produced.}
}
\value{
List with vector elements named with original column names.
}
\description{
All possible combinations of unique values of variables in a data frame (restricted to one value from each named column per combination).
}
