% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/UniqueValsFromVarnames_informatics_2dimensions.r
\name{UniqueValsFromVarnames}
\alias{UniqueValsFromVarnames}
\title{UniqueValsFromVarnames}
\usage{
UniqueValsFromVarnames(tb, varnames)
}
\arguments{
\item{tb}{Table input with named variables.}

\item{varnames}{vector of variable names from which unique values are to be extracted.}
}
\value{
List with vector elements named with original column names.
}
\description{
Create a list of unique values from named variables
}
