% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informatics1dimensional_UnlistVector.R
\name{UnlistVector}
\alias{UnlistVector}
\title{UnlistVector}
\usage{
UnlistVector(x)
}
\arguments{
\item{x}{Input object (must be table object with only one column)}
}
\value{
Vector object
}
\description{
Turns table with single column to into vector in one step
}
