% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarnamesOfClass_informatics_2dimensions.r
\name{VarnamesOfClass}
\alias{VarnamesOfClass}
\title{AsObject}
\usage{
VarnamesOfClass(
  tb,
  colclass = c("numeric", "character", "logical", "factor", "date", "integer")
)
}
\arguments{
\item{tb}{Table (data frame or tibble) whose names will be returned.}

\item{colclass}{Class of columns whose names will be returned. Must be one of the following: "numeric","character","logical","factor","date","integer".}
}
\value{
Character vector of names from tb matching colcass argument.
}
\description{
Output variable names in data frame which are of user-defined class
}
