% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoogleSheetLoadAllWorksheets_setup.r
\name{GoogleSheetLoadAllWorksheets}
\alias{GoogleSheetLoadAllWorksheets}
\title{GoogleSheetLoadAllWorksheets
Load all worksheets from a Google Sheet Object
Returns TRUE if expression throws an error, FALSE otherwise.}
\usage{
GoogleSheetLoadAllWorksheets(url)
}
\arguments{
\item{url}{Url of a Google Sheet}
}
\description{
GoogleSheetLoadAllWorksheets
Load all worksheets from a Google Sheet Object
Returns TRUE if expression throws an error, FALSE otherwise.
}
