% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumSubstringMatches_informatics.r
\name{NumSubstringMatches}
\alias{NumSubstringMatches}
\title{NumSubstringMatches}
\usage{
NumSubstringMatches(pattern, vector)
}
\arguments{
\item{pattern}{Pattern to be found as a substring with elements of the vector}

\item{vector}{Vector to be searched}
}
\value{
Numeric object with number of matches for 'pattern' within elements of 'vector'
}
\description{
Number of times specified substring occurs within vector of character strings
}
