% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informatics1dimensional_char.string.functions.r
\name{ReplaceRepeatingSubstringWithSingleSubstring}
\alias{ReplaceRepeatingSubstringWithSingleSubstring}
\title{ReplaceRepeatingSubstringWithSingleSubstring}
\usage{
ReplaceRepeatingSubstringWithSingleSubstring(string, substring)
}
\arguments{
\item{string}{Character string in which substrings are to be replaced}

\item{substring}{Repeated character pattern to be replaced with a single version of itself within 'string'}
}
\value{
'string' with no repeated substrings
}
\description{
Reduce repeating substring with single occurrence of that substring
}
